# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Onfly API",
        "description": "API de gestão de pedidos Onfly.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Ambiente local"
        }
    ],
    "paths": {
...
        "/pedido": {
            "get": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Listar pedidos de viagem",
                "description": "Lista os pedidos do usuário autenticado com filtros opcionais por status e período.",
                "operationId": "55228a76d283b8cb0fde57130eccbb09",
                "parameters": [
                    {
                        "name": "X-Request-Id",
                        "in": "header",
                        "description": "Identificador único da requisição (UUID).",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filtrar por status do pedido. 1=Solicitado, 2=Aprovado, 3=Cancelado.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "enum": [
                                1,
                                2,
                                3
                            ]
                        }
                    },
                    {
                        "name": "data_inicio",
                        "in": "query",
                        "description": "Data de início do filtro no formato DD-MM-YYYY.",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "01-01-2025"
                        }
                    },
                    {
                        "name": "data_fim",
                        "in": "query",
                        "description": "Data de fim do filtro no formato DD-MM-YYYY. Deve ser igual ou posterior a data_inicio.",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "31-01-2025"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de pedidos.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RespostaListaPedidos"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Header X-Request-Id ausente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RespostaErro"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RespostaErro"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Criar pedido de viagem",
                "description": "Cria um novo pedido de viagem para o usuário autenticado.",
                "operationId": "2503ac538d21340532cb460ed46b6964",
                "parameters": [
                    {
                        "name": "X-Request-Id",
                        "in": "header",
                        "description": "Identificador único da requisição (UUID).",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "solicitante",
                                    "data_partida",
                                    "data_retorno",
                                    "destino"
                                ],
                                "properties": {
                                    "solicitante": {
                                        "description": "Nome do solicitante da viagem.",
                                        "type": "string",
                                        "example": "João Silva"
                                    },
                                    "data_partida": {
                                        "description": "Data de partida no formato DD-MM-YYYY.",
                                        "type": "string",
                                        "example": "15-01-2025"
                                    },
                                    "data_retorno": {
                                        "description": "Data de retorno no formato DD-MM-YYYY. Deve ser igual ou posterior à data de partida.",
                                        "type": "string",
                                        "example": "20-01-2025"
                                    },
                                    "destino": {
                                        "description": "Dados do destino da viagem.",
                                        "properties": {
                                            "cidade": {
                                                "type": "string",
                                                "example": "São Paulo"
                                            },
                                            "estado": {
                                                "type": "string",
                                                "example": "SP"
                                            },
                                            "pais": {
                                                "type": "string",
                                                "example": "Brasil"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "status": {
                                        "description": "Status inicial do pedido. Opcional. 1=Solicitado, 2=Aprovado, 3=Cancelado.",
                                        "type": "integer",
                                        "example": 1,
                                        "enum": [
                                            1,
                                            2,
                                            3
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pedido criado com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RespostaPedido"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Header X-Request-Id ausente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RespostaErro"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RespostaErro"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Dados de entrada inválidos.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RespostaErro"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/pedido/{referencia_pedido}": {
            "get": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Buscar pedido por referência",
                "description": "Retorna os dados de um pedido específico pelo seu identificador único.",
                "operationId": "2802cae8eba6d4eab69040b62bef6094",
                "parameters": [
                    {
                        "name": "X-Request-Id",
                        "in": "header",
                        "description": "Identificador único da requisição (UUID).",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "referencia_pedido",
                        "in": "path",
                        "description": "Referência UUID do pedido.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pedido encontrado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RespostaPedido"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Header X-Request-Id ausente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RespostaErro"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RespostaErro"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pedido não encontrado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RespostaErro"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "RespostaErro": {
                "properties": {
                    "error": {
                        "type": "string",
                        "example": "Mensagem de erro."
                    }
                },
                "type": "object"
            },
...
            "Destino": {
                "properties": {
                    "cidade": {
                        "type": "string",
                        "example": "São Paulo"
                    },
                    "estado": {
                        "type": "string",
                        "example": "SP"
                    },
                    "pais": {
                        "type": "string",
                        "example": "Brasil"
                    }
                },
                "type": "object"
            },
            "Pedido": {
                "properties": {
                    "ref": {
                        "type": "string",
                        "format": "uuid",
                        "example": "9d4e3c2a-..."
                    },
                    "solicitante": {
                        "type": "string",
                        "example": "João Silva"
                    },
                    "data_partida": {
                        "description": "Formato DD-MM-YYYY",
                        "type": "string",
                        "example": "15-01-2025"
                    },
                    "data_retorno": {
                        "description": "Formato DD-MM-YYYY",
                        "type": "string",
                        "example": "20-01-2025"
                    },
                    "status": {
                        "type": "string",
                        "example": "Solicitado",
                        "enum": [
                            "Solicitado",
                            "Aprovado",
                            "Cancelado"
                        ]
                    },
                    "destino": {
                        "$ref": "#/components/schemas/Destino"
                    },
                    "criado_em": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-15T10:30:00.000000Z"
                    }
                },
                "type": "object"
            },
            "RespostaPedido": {
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/Pedido"
                    }
                },
                "type": "object"
            },
            "RespostaListaPedidos": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Pedido"
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token JWT obtido em POST /auth/token.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}